remove anonymous class creation to meet 5.4 version requirements
